# use lufact in inv of Symmetric/Hermitian matrices