Seems to be generally faster; with the following
```julia
using BenchmarkTools
luf(A) = Symmetric(inv(lufact(A)))
bkf(A) = Symmetric(inv(bkfact(A)))
for N in (10, 100, 500, 1000, 1500, 2000, 2500, 3000, 3500, 4000)
    A = rand(N, N); S = Symmetric(A);
    print("inv bkfact $N:"); @btime bkf($S)
    print("inv lufact $N:"); @btime luf($S)
end
```

I get:

| Matrix size | inv(bkfact(A))                  | inv(lufact(A))                  | ratio |
|:----------- |:------------------------------- |:------------------------------- |:----- |
| 10          | 5.467 μs (28% in `bkfact`)      | 2.219 μs (41% in `lufact`)      | 0.406 |
| 100         | 358.552 μs (25% in `bkfact`)    | 317.677 μs (43% in `lufact`)    | 0.886 |
| 500         | 8.246 ms (47% in `bkfact`)      | 8.669 ms (41% in `lufact`)      | 1.051 |
| 1000        | 48.386 ms (38% in `bkfact`)     | 52.142 ms (36% in `lufact`)     | 1.078 |
| 1500        | 200.098 ms (24% in `bkfact`)    | 156.982 ms (33% in `lufact`)    | 0.785 |
| 2000        | 502.729 ms (20% in `bkfact`)    | 370.605 ms (32% in `lufact`)    | 0.737 |
| 2500        | 1.016 s (18% in `bkfact`)       | 648.959 ms (32% in `lufact`)    | 0.639 |
| 3000        | 1.795 s (15% in `bkfact`)       | 1.165 s (31% in `lufact`)       | 0.649 |
| 3500        | 2.869 s (15% in `bkfact`)       | 1.795 s (23% in `lufact`)       | 0.626 |
| 4000        | 6.517 s (13% in `bkfact`)       | 2.818 s (29% in `lufact`)       | 0.432 |

Allocations are the same for both methods.

Would be nice if someone could confirm this on some other computer.