@tkelman 2017-07-05
how much is in factorize vs backsolve with this?