@fredrikekre 2017-07-05
Updated the table with percentage in factorization. It is definitely the backsolve that is slower for the Bunch Kaufman approach. Probably related to https://github.com/JuliaLang/julia/pull/22561#issuecomment-311360170?