@fredrikekre 2017-07-05
Updated to use the inplace `inv!` for LU, this cuts the allocations in half as well.