@fredrikekre 2017-07-05
Tested on my computer at home, which has better prestanda, here are the ratios:

| Matrix size | ratio |
|:----------- |:----- |
| 10          | 0.423 |
| 100         | 0.798 |
| 500         | 0.895 |
| 1000        | 1.140 |
| 1500        | 0.549 |
| 2000        | 0.385 |
| 2500        | 0.286 |
| 3000        | 0.254 |
| 3500        | 0.241 |
| 4000        | 0.236 |