@StefanKarpinski 2017-07-11
Shall we merge this then? The performance improvements look consistent and substantial.