@andreasnoack 2017-07-11
There is a funny little story about numerical analysis hidden here. I happen to have spend some time on this because of a small post to [NA Digest](http://www.netlib.org/na-digest-html/15/v15n28.html#2) in 2015. Suddenly Matlab's `inv` gave more inaccurate results than previous versions. After many experiments, I realized what was going on. If you use the LU to compute the inverse of an exactly symmetric matrix then the inverse is no longer symmetric.
```julia
julia> H = [1/(i + j - 1) for i in 1:8, j in 1:8];

julia> issymmetric(H)
true

julia> issymmetric(inv(lufact(H)))
false
```
This was the case in older versions of Matlab. It is pretty annoying that the inverse of a symmetric matrix is not symmetric and that was probably what made them change the behavior around R2013a. Suddenly, the matrix was symmetric. The way they did it was to use LU and then copy the elements from one triangle to another. That would effectively be the same as this PR. The funny thing is that
```julia
julia> norm(inv(H)*(H*ones(8)) - 1)
8.704438947487126e-7
```
is fine but
```julia
julia> norm(Symmetric(inv(lufact(H)))*(H*ones(8)) - 1)
0.2623294451513258
```
has a huge error. So the rounding errors in the LU based inverse somehow cancel accross the lower and upper triangle when solving. Also, this doesn't happen for Bunch-Kafman that only works on a single triangle
```julia
julia> norm(inv(bkfact(H))*(H*ones(8)) - 1)
7.277205904268713e-7
```
Matlab changed the behavior again in R2016a but with out mentioning anything in their release notes. A weird thing is that from and after 2016b I get different behavior on my Mac and on a Linux system. The Linux version is symmetric but the Mac inverse isn't. They both have small error now.

Bottomline is that we shouldn't merge this as it is.