@andreasnoack 2017-07-11
We could continue to use Bunch-Kaufman but since LU is so much faster it is probably better to figure out if symmetrizing with the average `(A+A')/2` (in place) is still faster. I would think so. The averaged version doesn't have the flaw
```julia
julia> inv(H) |> t -> Symmetric(t + t')/2*(H*ones(8)) - 1 |> norm
1.5769221751231655e-6
```