@fredrikekre 2017-07-11
Oh, thanks for catching that. This only applies to `inv` though, right? Things like `lufact(A)\b` when `A` is symmetric is okay I hope?