@fredrikekre 2017-07-11
> Yes. That is fine. The problem is only when copying elements from one triangle to the other in the unsymmetric LU based inverse.

Okay.

> We might actually want to add the Hilbert matrix example as a test case to make sure we don't accidentally make the mistake in the future.

Yea, planned to do that. For the in-place symmetrization, there is no such thing function implemented, right?