@andreasnoack 2017-07-11
> Yea, planned to do that. For the in-place symmetrization, there is no such thing function implemented, right?

Not to my knowledge