@fredrikekre 2017-07-11
Okay, I have an updated version now. Benchmarks:

| Matrix size | master (using bkfact) | PR (using lufact) | ratio (PR/master) |
|:----------- |:--------------------- |:----------------- |:----------------- |
| 10          | 5.236 μs              | 1.747 μs          | 0.334             |
| 100         | 203.389 μs            | 165.728 μs        | 0.815             |
| 500         | 4.635 ms              | 4.175 ms          | 0.901             |
| 1000        | 23.640 ms             | 27.766 ms         | 1.175             |
| 1500        | 165.135 ms            | 89.396 ms         | 0.541             |
| 2000        | 515.413 ms            | 214.738 ms        | 0.417             |
| 2500        | 1.208 s               | 365.067 ms        | 0.302             |
| 3000        | 2.312 s               | 627.485 ms        | 0.271             |
| 3500        | 3.867 s               | 998.406 ms        | 0.258             |
| 4000        | 5.969 s               | 1.548 s           | 0.259             |

and the allocations are still half for this PR.