@fredrikekre 2017-07-13
AV x86 failed with:
```
    FAILURE
Error in testset resolve:
Error During Test
  Test threw an exception of type ReadOnlyMemoryError
  Expression: sanity_tst(deps_data, problematic_data)
  ReadOnlyMemoryError()
  Stacktrace:
   [1] IOContext(::Base.AbstractIOBuffer{Array{UInt8,1}}, ::Pair{Symbol,Bool}) at .\show.jl:90 (repeats 2 times)
   [2] print(::Base.AbstractIOBuffer{Array{UInt8,1}}, ::Base.PipeServer) at .\strings\io.jl:29
   [3] print(::Base.AbstractIOBuffer{Array{UInt8,1}}, ::Base.PipeServer) at .\strings\io.jl:31
```

presumably unrelated?