@fredrikekre 2017-07-16
Noticed in #22827 that this enables `inv` for `Symmetric{BigFloat}` as well, since `lufact` works in that case, but there is no `BigFloat` implementation for `bkfact`! :tada: